Simplification pass: minimize drift from main, id-type-safe staleness
- SingletonState::singleton() goes back through StateManager::singleton()
  (matching main's shape) instead of inlining load(type, null)
- isStale() drops its int casts so staleness comparison works for
  ULID/UUID id types too (numeric for snowflakes, lexicographic-by-time
  for ULID/UUIDv7)—groundwork for the planned UUIDv7 move
- Upgrade guide: document the two new config keys

+++ src/State/ReconstitutingStateManager.php
@@ -224,9 +224,11 @@ class ReconstitutingStateManager extends StateManager
                 continue;
             }
 
-            $applied = $state->last_event_id ? (int) Id::from($state->last_event_id) : 0;
+            // No int casts: snowflake positions compare numerically either way,
+            // and ULID/UUIDv7 positions compare lexicographically-by-time.
+            $applied = $state->last_event_id ? Id::from($state->last_event_id) : 0;
 
-            if ((int) $max > $applied) {
+            if ($max > $applied) {
                 return true;
             }
         }
